+++ src/main/java/com/password4j/AlgorithmFinder.java
@@ -26,7 +26,6 @@ import java.security.SecureRandom;
 import java.security.Security;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.function.Function;
 
 /**
  * This utility class finds algorithms with their configuration
@@ -143,7 +142,8 @@ public class AlgorithmFinder
      */
     public static PBKDF2Function getPBKDF2Instance()
     {
-        return getPBKDF2Instance(a -> (i -> l -> (PBKDF2Function.getInstance(a, i, l))));
+        Param params = internalGetProperties();
+        return PBKDF2Function.getInstance(params.algorithm, params.iterations, params.length);
     }
 
     /**
@@ -181,15 +181,16 @@ public class AlgorithmFinder
      */
     public static CompressedPBKDF2Function getCompressedPBKDF2Instance()
     {
-        return getPBKDF2Instance(a -> (i -> l -> (CompressedPBKDF2Function.getInstance(a, i, l))));
+        Param params = internalGetProperties();
+        return CompressedPBKDF2Function.getInstance(params.algorithm, params.iterations, params.length);
     }
 
-    private static <T extends PBKDF2Function> T getPBKDF2Instance(Function<String, Function<Integer, Function<Integer, T>>> f)
+    private static Param internalGetProperties()
     {
         String algorithm = PropertyReader.readString("hash.pbkdf2.algorithm", Hmac.SHA512.name(), "PBKDF2 algorithm is not defined");
         int iterations = PropertyReader.readInt("hash.pbkdf2.iterations", 64_000, "PBKDF2 #iterations are not defined");
         int length = PropertyReader.readInt("hash.pbkdf2.length", Hmac.SHA512.bits(), "PBKDF2 key length is not defined");
-        return f.apply(algorithm).apply(iterations).apply(length);
+        return new Param(algorithm, iterations, length);
     }
 
     /**
@@ -206,6 +207,11 @@ public class AlgorithmFinder
      *     <th>Default</th>
      *   </tr>
      *   <tr>
+     *      *     <td>Logarithmic number of rounds</td>
+     *      *     <td>hash.bcrypt.minor</td>
+     *      *     <td>10</td>
+     *      *   </tr>
+     *   <tr>
      *     <td>Logarithmic number of rounds</td>
      *     <td>hash.bcrypt.rounds</td>
      *     <td>10</td>
@@ -217,8 +223,9 @@ public class AlgorithmFinder
      */
     public static BCryptFunction getBCryptInstance()
     {
+        char minor = PropertyReader.readChar("hash.bcrypt.minor", 'b', "BCrypt minor version is defined");
         int rounds = PropertyReader.readInt("hash.bcrypt.rounds", 10, "BCrypt rounds are not defined");
-        return BCryptFunction.getInstance(rounds);
+        return BCryptFunction.getInstance(BCrypt.valueOf(minor), rounds);
     }
 
     /**
@@ -288,5 +295,19 @@ public class AlgorithmFinder
     private static boolean useStrongRandom()
     {
         return PropertyReader.readBoolean("global.random.strong", false);
+    }
+
+    private static class Param
+    {
+        String algorithm;
+        int iterations;
+        int length;
+
+        Param(String algorithm, int iterations, int length)
+        {
+            this.algorithm = algorithm;
+            this.iterations = iterations;
+            this.length = length;
+        }
     }
 }